To check whether the given number is prime or not!
Tried by using if statement initially. But then output was wrong. Glanced the solution and saw that they had used for loop. I then did the same.
However they have not made the program complex by adding the numbers 1 & 2 as input. I tried to do that but even if the 'if' statement gets satisfied, still it enters the for loop which is giving output two times. So :(
Good try still :)